refactor: :label: refactor type definitions files. move capitalizeFirstLetter and romanize into a  stringUtils file
