--- src/utils/capitalizeFirstLetter.tsx
@@ -1,2 +1,0 @@
-export const capitalizeFirstLetter = (string: string) =>
-  string.charAt(0).toUpperCase() + string.slice(1);
