+++ src/utils/stringUtils.ts
@@ -12,3 +12,5 @@ const map = [
 ] as const;
 
 export const romanize = (number: number) => map[number - 1];
+export const capitalizeFirstLetter = (string: string) =>
+  string.charAt(0).toUpperCase() + string.slice(1);
